Update: Move SyntaxTree to ast-node-factory.js
